Add micros, millis, seconds, systicks, and sysoverflows
```micros``` and ```millis``` are standard Arduino functions that return the time in microseconds and milliseconds since the program has started. To acheive this we are using the Apollo3's stimer peripheral. This peripheral is a non-settable (though you can 'clear' it to zero) 32-bit timer that retains its value through anything except a full power-on-reset. We ensure that the value is cleared in ```ap3_init``` so that the value accurately represents the time since the program has begun. Using a 3MHz source this timer overflows in ~23 minutes, at which time the overflow interrupt is responded to. Another 32-bit variable tracks the higher-order bits of the stimer in software. The combined 64-bit value will last ~195,000 years before overflowing.

As an extension to Arduino's interface users may now also access ```seconds```, ```systicks```, and ```sysoverflows```. These functions respectively return elapsed seconds from program start, the value of the 32-bit stimer counter, and the 32-bit value tracking the number of times that the stimer counter has overflowed since program execution began.