Provide 'end' symbol for C std library
This replaces the previous definition of 'end' that was in error (it was an empty void void function). This is a (more) appropriate solution that provides a memory address that is computed from the linker script. 'end' allows malloc + friends to operate correctly. This in turn allows some other important functions to work, like Arduino Strings and stdlib 'rand()'