Increasing wrong character timeout to 30 to fix 'no annoucement' error.
I had a sketch reporting 'No announcement from Artemis bootloader'. This occurs when either 1) no chr(127) is received or 2) too many non-chr(127) characters are received. The sketch I was working on transmits a ton of serial data at 115200 constantly. After the Artemis is reset, the python script has more than 10 bytes sitting in the buffer. 30 seems to work well. With a 0.1s timeout, this is 3s total. Increasing to 50 seems overkill.