Add 'wait_for_pulse' to ap3_pwm_output
This waits until a pulse has just been completed to change the compare register values. This attempts to solve isse #16.

servoWrite could cause the signal to be inverted if/when the compare registers were changed at the same time that they matched the timer. This attempts to wait until the timer value is below the smaller compare register to make any changes.