Add SBL baud rate menu choice for Edge
Now Edge board users have two different baud rate options via the menus. One is used for the Ambiq Secure Bootloader (SBL) and the other is used for the SparkFun Variable Loader (SVL).

Two baud rates are provided - 921600 for the majority of boards out there which were purchased from SparkFun, and 115200 for the TensorFlow conference boards.

Edge boards do not come with the SVL, so they default to using the SBL at 921600. You may use SBL to load the SVL sketch and thus get the benefits of the variable bootloader.

Future improvements would be to hide the baud rate selection for the bootloader that is not selected, but this only affects the Edge variant.