Shift burden of baud negotiation to target board
To reduce the impact of differences between operating systems and USB-serial drivers on the bootloading process we've shifted the baud rate negotiation onto the target board. The host computer is able to begin communication with the desired settings and the target tries a list of baud rates until either a match is found or all rates have failed. Once the baud rate is found bootloading proceeds identically to v1.