Use 'attachInterruptArg()' for RX interrupts with 'this' argument
This allows us to dispose of the ugly global table of pointers to  software serial objects. Instead those pointers are all handled in the GPIO core.

We will need to make sure to detach the interrupt (and thereby remove reference to the software serial object) when the object is destroyed.