Removed code using <time.h> non-core library (ie. Epoch stuff)
I hate to undo even a little bit of Adam's good work, but in the interests of a robust core, I think it's best the references to time.h are removed, and the Epoch set/get along with it.

RTC.cpp included <time.h>, which caused a compilation error when these are also included in code:
#include <Time.h>
#include <TimeLib.h>

Some possible issues with Window lack of case sensitivity, refer here: https://forum.arduino.cc/index.php?topic=451360.0

Also with time.h (or Time.h?) being an optional extra install library to the Arduino IDE, best to leave it out of the core, especially if it can cause compilation errors.