patch ble
it appears that simply changing the mbed libs used by the variants back to the previous release is enough to resolve the BLE 'starting failed' issue.
this suggests that some change in our mbed port is the root of the issue and needs further investigation.
it also suggests that the micromod artemis processor board might not support ble until an mbed lib is generated with the proper version of mbed